+++ src/hash-string.c
@@ -24,11 +24,11 @@ CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
 /* String hash function */
 
-unsigned long string_hash(void *string)
+unsigned int string_hash(void *string)
 {
 	/* This is the djb2 string hash function */
 
-	unsigned long result = 5381;
+	unsigned int result = 5381;
 	unsigned char *p;
 
 	p = (unsigned char *) string;
@@ -44,9 +44,9 @@ unsigned long string_hash(void *string)
 /* The same function, with a tolower on every character so that 
  * case is ignored.  This code is duplicated for performance. */
 
-unsigned long string_nocase_hash(void *string)
+unsigned int string_nocase_hash(void *string)
 {
-	unsigned long result = 5381;
+	unsigned int result = 5381;
 	unsigned char *p;
 
 	p = (unsigned char *) string;
